refactor EnrollmentPage into separate tab components
